Fix Linux compile bustage Review URL: http://codereview.chromium.org/119378
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17972 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/thumbnail_store.cc
@@ -89,7 +89,7 @@ bool ThumbnailStore::GetPageThumbnailFromDisk(const FilePath& file,
   int out_len;
 
   if (!packed.ReadData(&iter, &jpeg_data, &out_len) ||
-      out_len != jpeg_len)
+      out_len != static_cast<int>(jpeg_len))
     return false;
 
   // Copy jpeg data to the out parameter.
